requirements.txt
fastapi[all]
playwright
requests
beautifulsoup4
lxml
uvicorn
undetected-chromedriver
selenium
mysql-connector-python
crawl4ai
bcrypt